Commit for loan and payment operation

+++ src/main/java/com/aplazo/mx/creditlines/business/CreditLineCalculatorStrategy.java
@@ -3,5 +3,5 @@ import java.math.BigDecimal;
 import java.util.Optional;
 
 public interface CreditLineCalculatorStrategy {
-    Optional<BigDecimal> calculateCreditLine(int age);
+    Optional<Double> calculateCreditLine(int age);
 }
